requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database
psycopg2-binary==2.9.10
supabase==2.0.2

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7

# Image handling
Pillow==11.3.0
django-imagekit==5.0.0

# Payment gateways
razorpay==1.4.2
stripe==7.8.0

# Celery for async tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Email
django-ses==3.5.2

# Forms & UI
django-crispy-forms==2.1
crispy-bootstrap4==2022.1

# Utilities
python-decouple==3.8
django-extensions==3.2.3
django-debug-toolbar==4.2.0
django-seed==0.3.1

# API Documentation
drf-spectacular==0.26.5

# Export/Import
django-import-export==3.3.1
openpyxl==3.1.2
reportlab==4.0.7

# Caching
django-redis==5.4.0

# Testing
factory-boy==3.3.0
pytest-django==4.7.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
sentry-sdk[django]==1.38.0